pyproject.toml
[project]
name = "trader-bro"
version = "0.1.0"
description = "Claude-powered crypto trading bot"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]",
    "uvicorn[standard]",
    "anthropic",
    "requests",
    "pyjwt[crypto]",
    "python-dotenv",
]

[tool.hatch.build.targets.wheel]
packages = ["bot"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
